backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
sqlalchemy==1.4.42
databases[postgresql]==0.8.0
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.2
alembic==1.13.0
